Fix parsing of platform specific sections in "Functions" gamedata
The "Functions" section gamedata parser didn't skip the event of leaving an ignored sub section (like a "windows" section on a linux host) correctly, but counted it as leaving the parent section instead. This corrupted the parser state and made it parse the following entries in the wrong level.

Thanks vanz.