Fix trying to call hook removal callback on unloaded plugin
When a plugin which registered an entity vtable hook with a remove callback is unloaded OnMapEnd, dhooks still tried to call the removal callback on the next frame after the map ended and all entities were destroyed.

Remove the plugin's hooks from the removal list when a plugin is unloaded. The hooks are free'd anyways.

Drifter321/DHooks2#3