Add methodmap API
Vtable hooks are wrapped in a `DynamicHook` methodmap and detours in a `DynamicDetour` methodmap. Parameters and return values pushed to the callbacks are of methodmap types `DHookParam` and `DHookReturn` to easily access the values.

Most functions are mapped 1:1 with the exception of creating vtable hooks, where the callback function has to be passed in the hook function instead of during setup.

`HookRaw` dropped the `removecb` parameter, since it was never called. There currently is no way to know when to unhook the raw object.

This change doesn't affect the legacy API at all.